Combine membership forms
- Move logic into single page and composables
- Use pattern library content card for containers
- Delete unused components and composables

Ticket: https://phabricator.wikimedia.org/T396600
