+++ src/scss/pattern-library-compatibility/section-heading.scss
@@ -6,4 +6,8 @@
 	h2 {
 		font-size: 20px;
 	}
+
+	hr {
+		margin: 16px 0;
+	}
 }
